# java-projects/api--bff/src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: api-bff API
  description: BFF API.
  version: 0.1.9
servers:
  - url: http://localhost:0
    description: Optional server description, e.g. Main (production) server
paths:
...
  /brokers/{id}:
    get:
      summary: Returns a broker by the Id.
      operationId: getBrokerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200': # status code
          description: A JSON object of a broker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Broker'
components:
  schemas:
...
    Broker:
      type: object
      required:
        - brokerShortName
        - brokerName
        - brokerCnpj
      properties:
        id:
          type: integer
          format: int32
        brokerShortName:
          type: string
        brokerName:
          type: string
        brokerCnpj:
          type: string